使用redis中的hincrby进行未读消息+1，减少复杂度

+++ xigua-domain/src/main/java/com/xigua/domain/vo/LastChatVO.java
@@ -20,5 +20,5 @@ public class LastChatVO implements Serializable {
     private String avatar;
     private String lastMessage;
     private Boolean isOnline;
-    private Integer friendUnreadCount;
+    private Long friendUnreadCount;
 }
